# back/transport-api/docs/swagger.yaml
basePath: /api/v1
definitions:
  ads.adData:
    properties:
      ad_id:
        type: integer
      company_id:
        type: integer
      created_at:
        type: integer
      created_id:
        type: integer
      industry:
        type: string
      price:
        type: integer
      responsible_id:
        type: integer
      text:
        type: string
      title:
        type: string
    type: object
  ads.adListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/ads.adData'
        type: array
      error:
        type: string
    type: object
  ads.adResponse:
    properties:
      data:
        $ref: '#/definitions/ads.adData'
      error:
        type: string
    type: object
  ads.addAdRequest:
    properties:
      industry:
        type: string
      price:
        type: integer
      text:
        type: string
      title:
        type: string
    type: object
  ads.updateAdRequest:
    properties:
      price:
        type: integer
      text:
        type: string
    type: object
  companies.companyData:
    properties:
      created_at:
        type: integer
      industry:
        type: string
      owner_id:
        type: integer
      title:
        type: string
    type: object
  companies.companyResponse:
    properties:
      data:
        $ref: '#/definitions/companies.companyData'
      error:
        type: string
    type: object
  companies.industryResponse:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
  companies.mainPageData:
    properties:
      stats:
        $ref: '#/definitions/companies.mainPageStatsData'
      title:
        type: string
    type: object
  companies.mainPageResponse:
    properties:
      data:
        $ref: '#/definitions/companies.mainPageData'
      error:
        type: string
    type: object
  companies.mainPageStatsData:
    properties:
      active_ads_aount:
        type: integer
      active_leads_amount:
        type: integer
      active_leads_price:
        type: integer
      closed_leads_price:
        type: integer
      company_absolute_rating:
        type: number
      company_relative_rating:
        type: number
      total_leads_amount:
        type: integer
    type: object
  companies.updateCompanyRequest:
    properties:
      industry:
        type: integer
      owner_id:
        type: integer
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Swagger документация к API
  title: BRM API
  version: "1.0"
paths:
  /ads:
    get:
      description: Получает список объявлений с использованием фильтрации и пагинации
      parameters:
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      - description: Поиск по названию/тексту
        in: query
        name: name
        required: true
        type: string
      - description: Поиск по отрасли
        in: query
        name: industry
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение объявлений
          schema:
            $ref: '#/definitions/ads.adListResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/ads.adListResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/ads.adListResponse'
      summary: Получение списка объявлений
      tags:
      - ads
    post:
      consumes:
      - application/json
      description: Добавляет новое объявление
      parameters:
      - description: Новое объявление в JSON
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ads.addAdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное добавление объявления
          schema:
            $ref: '#/definitions/ads.adResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/ads.adResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/ads.adResponse'
      summary: Добавление нового объявления
      tags:
      - ads
  /ads/{id}:
    delete:
      description: Безвозвратно удаляет объявление
      parameters:
      - description: id объявления
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление объявления
          schema:
            $ref: '#/definitions/ads.adResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/ads.adResponse'
        "404":
          description: Объявление не найдено
          schema:
            $ref: '#/definitions/ads.adResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/ads.adResponse'
      summary: Удаление объявления
      tags:
      - ads
    get:
      description: Получает объявление по id
      parameters:
      - description: id объявления
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение объявления
          schema:
            $ref: '#/definitions/ads.adResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/ads.adResponse'
        "404":
          description: Объявление не найдено
          schema:
            $ref: '#/definitions/ads.adResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/ads.adResponse'
      summary: Получение объявления
      tags:
      - ads
    put:
      consumes:
      - application/json
      description: Изменяет одно или несколько полей объявления
      parameters:
      - description: id объявления
        in: path
        name: id
        required: true
        type: integer
      - description: Новые поля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ads.updateAdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление данных
          schema:
            $ref: '#/definitions/ads.adResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/ads.adResponse'
        "404":
          description: Объявление не найдено
          schema:
            $ref: '#/definitions/ads.adResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/ads.adResponse'
      summary: Редактирование объявления
      tags:
      - ads
  /companies/{id}:
    delete:
      description: Безвозвратно удаляет компанию и всё, что с этой компанией связано
      parameters:
      - description: id компании
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление компании
          schema:
            $ref: '#/definitions/companies.companyResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/companies.companyResponse'
        "404":
          description: Компания не найдена
          schema:
            $ref: '#/definitions/companies.companyResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/companies.companyResponse'
      summary: Удаление компании
      tags:
      - core/companies
    get:
      description: Возвращает поля компании для страницы редактирования
      parameters:
      - description: id компании
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение данных
          schema:
            $ref: '#/definitions/companies.companyResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/companies.companyResponse'
        "404":
          description: Компания не найдена
          schema:
            $ref: '#/definitions/companies.companyResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/companies.companyResponse'
      summary: Получение информации о компании
      tags:
      - core/companies
    put:
      description: Изменяет одно или несколько полей компании
      parameters:
      - description: id компании
        in: path
        name: id
        required: true
        type: integer
      - description: Новые поля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/companies.updateCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление данных
          schema:
            $ref: '#/definitions/companies.companyResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/companies.companyResponse'
        "404":
          description: Компания не найдена
          schema:
            $ref: '#/definitions/companies.companyResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/companies.companyResponse'
      summary: Редактирование полей компании
      tags:
      - core/companies
  /companies/industries:
    get:
      description: Возвращает словарь из отраслей и их id
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение данных
          schema:
            $ref: '#/definitions/companies.industryResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/companies.industryResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/companies.industryResponse'
      summary: Получение отраслей
      tags:
      - core/companies
  /companies/mainpage/{id}:
    get:
      description: Возвращает название и статистику компании для главной страницы
      parameters:
      - description: id компании
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение данных
          schema:
            $ref: '#/definitions/companies.mainPageResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/companies.mainPageResponse'
        "404":
          description: Компания не найдена
          schema:
            $ref: '#/definitions/companies.mainPageResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/companies.mainPageResponse'
      summary: Получение информации о компании
      tags:
      - core/companies
...
swagger: "2.0"